Rename concludable package to resolvable
